# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TugrikRate Rates API",
    "description": "Exchange rates for Mongolian tugrik (MNT), collected from 15 Mongolian banks and financial institutions.\n\n`GET /v1/rates` is the stable public contract. Rates are JSON **strings** so exact decimal precision survives the wire, and each quote states the channel and side the source actually publishes - nothing is copied between channels or inferred.",
    "contact": {
      "name": "Battseren Badral (upstream)",
      "url": "https://github.com/btseee/mongolian-bank-exchange-rate"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/v1/rates": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "All sources, latest rates",
        "description": "Every source's most recent rates.\n\nSupports `If-None-Match`: an app polling on an interval gets a bare\n304 when nothing has moved, which matters on a metered connection.",
        "operationId": "get_rates_v1_rates_get",
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated ISO codes, e.g. USD,EUR",
              "examples": [
                "USD,EUR"
              ],
              "title": "Currency"
            },
            "description": "Comma-separated ISO codes, e.g. USD,EUR"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated source ids",
              "examples": [
                "khanbank"
              ],
              "title": "Source"
            },
            "description": "Comma-separated source ids"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatesResponse"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Send back as If-None-Match.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Unchanged since the supplied ETag; no body."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fx": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Foreign-exchange table: units of each currency per 1 USD",
        "description": "International rates for converting one **foreign** currency into\nanother, with no MNT and no Mongolian bank involved.\n\nEvery rate is how many units of `currency` one US dollar buys.\nUSD to X is `rate`; X to USD is `1 / rate`; X to Y (neither USD) is\n`rate_Y / rate_X`, using both rates from the **same source** so they\ncome from one snapshot. The same table supplies the USD leg when a\nMongolian rate is missing - see \"Conversion policy\" in\nARCHITECTURE.md. Refreshed a few times a day, not every 15 minutes.",
        "operationId": "get_fx_v1_fx_get",
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated ISO codes, e.g. KZT,EUR",
              "examples": [
                "KZT,EUR"
              ],
              "title": "Currency"
            },
            "description": "Comma-separated ISO codes, e.g. KZT,EUR"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated source ids",
              "examples": [
                "frankfurter"
              ],
              "title": "Source"
            },
            "description": "Comma-separated source ids"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FxResponse"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Send back as If-None-Match.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Unchanged since the supplied ETag; no body."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rates/{source_id}/history": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Snapshot history for one source",
        "description": "Distinct published rate sets for one source, newest first.\n\nOne entry per actual change - repeated crawls that found the same\npayload do not create entries, they move `last_checked_at`.",
        "operationId": "get_history_v1_rates__source_id__history_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown source id."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FxResponse": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "description": "Contract version. Bumped on any breaking change.",
            "examples": [
              1
            ]
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "base": {
            "type": "string",
            "title": "Base",
            "description": "Every rate is quoted against this currency.",
            "examples": [
              "USD"
            ]
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/FxSourceOut"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "base",
          "sources"
        ],
        "title": "FxResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryResponse": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "source": {
            "$ref": "#/components/schemas/HistorySourceOut"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/HistorySnapshotOut"
            },
            "type": "array",
            "title": "Snapshots"
          }
        },
        "type": "object",
        "required": [
          "schema_version",
          "source",
          "snapshots"
        ],
        "title": "HistoryResponse"
      },
...
      "HistorySourceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_mn": {
            "type": "string",
            "title": "Name Mn"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "name_mn",
          "type"
        ],
        "title": "HistorySourceOut"
      },
...
      "RatesResponse": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "description": "Contract version. Bumped on any breaking change.",
            "examples": [
              1
            ]
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOut"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "schema_version",
          "generated_at",
          "sources"
        ],
        "title": "RatesResponse"
      },
...
      "SourceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "khanbank"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Official English name.",
            "examples": [
              "Khan Bank"
            ]
          },
          "name_mn": {
            "type": "string",
            "title": "Name Mn",
            "description": "Official Mongolian (Cyrillic) name.",
            "examples": [
              "Хаан Банк"
            ]
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url",
            "description": "Absolute URL of this source's logo (PNG or JPEG, square where the source has an app icon). Null when no logo of verified provenance exists. The URL embeds a content hash, so it can be cached indefinitely.",
            "examples": [
              "https://api.example.com/static/logos/khanbank.jpg?v=ab12cd34ef56"
            ]
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "commercial_bank | central_bank | exchange_bureau | remittance | international_aggregator. Treat as non-exhaustive.",
            "examples": [
              "commercial_bank"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "ok | stale | failing. A stale or failing source still returns its last good quotes.",
            "examples": [
              "ok"
            ]
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At",
            "description": "When this service retrieved these rates."
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At",
            "description": "When the source says it published them. Null when the source does not say - never inferred from fetch time."
          },
          "last_checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At",
            "description": "Most recent crawl that returned this same payload."
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteOut"
            },
            "type": "array",
            "title": "Quotes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "name_mn",
          "logo_url",
          "type",
          "status",
          "fetched_at",
          "published_at",
          "last_checked_at",
          "quotes"
        ],
        "title": "SourceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    }
  },
...
}